Extract AbiPlaygroundLink into a component
